Guard `where_sql` call with a better check
The `:ast` method was added in Arel v2.0.x, and we currently only
support >= v7. The `arel` argument can be both an `::Arel::TreeManager`
instance or a `String`, and probably it could be other objects as well.
But checking whether it responds to `:where_sql` is more explicit than
checking for `:ast`.